Relocate pear errors
