+++ src/commands/current/index.ts
@@ -1,7 +1,7 @@
 import { Command } from "@oclif/command"
 
 import { PearData } from "../../pear-data"
-import { PearError } from "../../pear-errors"
+import { PearError } from "../../shared/PearErrors"
 
 export default class Current extends Command {
   static description = "list current authors"
