+++ src/commands/known/add.ts
@@ -1,8 +1,8 @@
 import { Command } from "@oclif/command"
 
 import { PearData } from "../../pear-data"
-import { PearError } from "../../pear-errors"
 import { Pear } from "../../shared/Pear"
+import { PearError } from "../../shared/PearErrors"
 
 export default class AddKnown extends Command {
   static description = "add known author"
