+++ src/shared/Pear.ts
@@ -1,7 +1,9 @@
+import { PearErrors } from "./PearErrors"
 import { PearMessages } from "./PearMessages"
 import { PearUtils } from "./PearUtils"
 
 export const Pear = {
+  errors: PearErrors,
   messages: PearMessages,
   utils: PearUtils,
 }
